+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## [3.2.0] 2021-12-31
 
+### ADDED
+
+- Added option to recover hit dice before rolling.
+
 ### CHANGED
 
 - Bump core compatible version to 0.9.
+- Auto-roll hit dice when recovery fraction is set to "Full".
+
 ## [3.1.0] 2021-06-19
 
 ### ADDED